finish up on shifts having orders and tasks instead of year,month,day; better SQLAlchemy code for getting an object via primary key, cascade delete workgroup->tasks, Unique constraint for tasks (label, wg_id), various little fixes
git-svn-id: http://subversion.assembla.com/svn/vokomokum/trunk@94 b3c4d6f4-6639-42c7-84c2-d72431877607

+++ members/members/views/workgroup.py
@@ -23,7 +23,7 @@ def mkworkgroup(session, request=None, wg_id=None):
             request.matchdict['wg_id'] != 'fresh') or wg_id:
         if not wg_id:
             wg_id = request.matchdict['wg_id']
-        wg = session.query(Workgroup).filter(Workgroup.id==wg_id).first()
+        wg = session.query(Workgroup).get(wg_id)
         if wg:
             wg.exists = True
     else:
@@ -79,7 +79,7 @@ class WorkgroupView(BaseView):
         order_id = self.request.params.has_key('order_id') and int(self.request.params['order_id'])\
                         or list(order_header)[0].ord_no
 
-        self.order = session.query(Order).filter(Order.id==order_id).first()
+        self.order = session.query(Order).get(order_id)
         self.orders = session.query(Order).all()
 
         shifts = session.query(Shift).filter(Shift.wg_id==wg.id)\
@@ -121,51 +121,57 @@ class WorkgroupEditView(BaseView):
                 if self.request.params.has_key('wg_members'):
                     wg.members = []
                     for mid in self.request.POST.getall('wg_members'):
-                        m = session.query(Member).filter(Member.id == mid).first()
+                        m = session.query(Member).get(mid)
                         wg.members.append(m)
                 try:
                     session.add(wg)
                     session.flush()
                     new_id = wg.id
                     transaction.commit()
-                    session = DBSession()
                 except Exception, e:
                     return dict(wg = None, msg=u'Something went wrong: %s' % e)
                 self.possible_members = get_possible_members(session)
                 return dict(wg = mkworkgroup(session, self.request, wg_id=new_id), msg='Workgroup has been saved.')
 
             elif action == 'delete':
+                wg_name = wg.name
                 try:
-                    session.delete(session.query(Workgroup)\
-                        .filter(Workgroup.id == wg.id).first()\
-                    )
+                    session.delete(session.query(Workgroup).get(wg.id))
                     session.flush()
                     transaction.commit()
                 except Exception, e:
                     return dict(wg = None, msg=u'Something went wrong: %s' % e)
-                return dict(wg = None, msg='Workgroup %s has been deleted.' % wg)
+                return dict(wg = None, msg='Workgroup %s has been deleted.' % wg.name)
 
             elif "task" in action:
-                print action, "11111111112"
+                print action, " =========================================================================="
                 msg = ''
                 try:
-                    if action == "add-task":
+                    if action == 'add-task':
                         task = Task(self.request.params['task_label'], wg.id)
+                        wg.tasks.append(task)
                         session.add(task)
+                        msg = 'Added task.'
                     elif action == 'toggle-task-activity':
-                        task = session.query(Task).filter(Task.id==self.request.params['task_id']).first()
+                        task = session.query(Task).get(self.request.params['task_id'])
                         task.active = not task.active
                         session.add(task)
+                        msg = 'changed activity status of the task.'
                     elif action == 'delete-task':
-                        task = session.query(Task).filter(Task.id==self.request.params['task_id']).first()
-                        session.delete(task)
+                        task = session.query(Task).get(self.request.params['task_id'])
+                        shifts = session.query(Shift).filter(Shift.task_id==task.id).all()
+                        if len(shifts) == 0:
+                            session.delete(task)
+                            msg = 'Deleted task.'
+                        else:
+                            msg = 'Cannot delete task, as there are shifts in the history with this task.'
                     session.flush()
                     transaction.commit()
                 except Exception, e:
                     #TODO: redirect to custom 404 view?
                     msg = ' Could not perform %s: %s' % (action, e)
                     #TODO: re-initiating this too often (after each commit), there must be a more elegant way
-                    self.possible_members = get_possible_members(session)
+                self.possible_members = get_possible_members(session)
                 return dict(wg = mkworkgroup(session, self.request), msg = msg)
